pyproject.toml
[tool.poetry]
name = "r.umbra"
version = "0.1.0"
description = "A general Discord bot."
authors = ["AbstractUmbra <Umbra@AbstractUmbra.xyz>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = { git = "https://github.com/Rapptz/discord.py", rev = "master" }
discord-ext-menus = { git = "https://github.com/Rapptz/discord-ext-menus", rev = "master" }
jishaku = { git = "https://github.com/Gorialis/jishaku", rev = "master" }
wavelink = "*"
"mystbin.py" = "*"
"discord.py-stubs" = "^1.5.1"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
